minor changes to app.js. Created header file and modified nav.js. moved login and signup into pages folder
