+++ client/src/components/Header.js
@@ -1,0 +1,14 @@
+import React, { Component } from "react";
+
+class Header extends Component {
+    render() {
+        return (
+            <header>
+            <h1>Gourmet-Recipes</h1>  
+            <p>Delicious recipes at the tip of your fingers</p>
+            </header> 
+        );
+    }
+}
+
+export default Header;
