+++ client/src/components/Nav.js
@@ -1,14 +1,39 @@
 import React from "react";
 import {Link} from "react-router-dom";
+import {MDBBtn} from "mdbreact";
 
 function Nav() {
-    return(
-        <nav>
-            <h2>Gourmet-Recipes</h2>
-            <Link to="/">Logout</Link>
-            <Link to="/login">Login</Link>
-        </nav>
-    );
+  return (
+    <nav>
+      <ul>
+        <li>
+          <MDBBtn href="/home">Log Out</MDBBtn>
+        </li>
+        <br></br>
+        <li>
+          <MDBBtn href="/login">Login</MDBBtn>
+        </li>
+        <br></br>
+        <li>
+          <MDBBtn href="/signup">Sign Up</MDBBtn>
+        </li>
+        <br></br>
+        <li>
+          <Link
+            className="link"
+            activeClass="active"
+            to="contacts"
+            spy={true}
+            smooth={true}
+            offset={-70}
+            duration={500}
+          >
+            Contact Us
+          </Link>
+        </li>
+      </ul>
+    </nav>
+  );
 }
 
 export default Nav;
